add qat test and refine code
