+++ onnxruntime/python/tools/quantization/quant_utils.py
@@ -4,9 +4,10 @@ from enum import Enum
 
 from pathlib import Path
 
+__producer__ = "onnx.quantize"
+__version__ = "0.1.0"
 onnx_domain = "ai.onnx"
 ms_domain = "com.microsoft"
-onnx_op_set_version = 11
 
 
 type_to_name = {
